Revert "change type to int"
This reverts commit d3de2d0a9afbae661cd2fee6f0278d74880cbee2.

+++ src/main/kotlin/Summary.kt
@@ -1,5 +1,5 @@
 data class Summary(
-    val income: Int,
-    val expenses: Int,
-    val remaining: Int
+    val income: Double,
+    val expenses: Double,
+    val remaining: Double
 )
